refactor: 할당 생성에 대한 validator 클래스 분리 및 findBy 대신 existsBy 사용
